cubecl::channel

Trait ComputeChannel

pub trait ComputeChannel<Server>:
    Clone
    + Debug
    + Send
    + Sync
where
    Server: ComputeServer,
{
    // Required methods
    fn read(
        &self,
        bindings: Vec<Binding>,
    ) -> impl Future<Output = Vec<Vec<u8>>> + Send;
    fn get_resource(&self, binding: Binding) -> BindingResource<Server>;
    fn create(&self, data: &[u8]) -> Handle;
    fn empty(&self, size: usize) -> Handle;
    unsafe fn execute(
        &self,
        kernel: <Server as ComputeServer>::Kernel,
        count: CubeCount,
        bindings: Vec<Binding>,
        mode: ExecutionMode,
    );
    fn flush(&self);
    fn sync(&self) -> impl Future<Output = ()> + Send;
    fn sync_elapsed(
        &self,
    ) -> impl Future<Output = Result<Duration, TimestampsError>> + Send;
    fn memory_usage(&self) -> MemoryUsage;
    fn enable_timestamps(&self);
    fn disable_timestamps(&self);
}

The ComputeChannel trait links the ComputeClient to the ComputeServer while ensuring thread-safety

Required Methods

fn read( &self, bindings: Vec<Binding>, ) -> impl Future<Output = Vec<Vec<u8>>> + Send

Given bindings, returns owned resources as bytes

fn get_resource(&self, binding: Binding) -> BindingResource<Server>

Given a resource handle, return the storage resource.

fn create(&self, data: &[u8]) -> Handle

Given a resource as bytes, stores it and returns the resource handle

fn empty(&self, size: usize) -> Handle

Reserves size bytes in the storage, and returns a handle over them

unsafe fn execute( &self, kernel: <Server as ComputeServer>::Kernel, count: CubeCount, bindings: Vec<Binding>, mode: ExecutionMode, )

Executes the kernel over the given bindings.

Safety

When executing with mode ExecutionMode::Unchecked, out-of-bound reads and writes can happen.

fn flush(&self)

Flush outstanding work of the server.

fn sync(&self) -> impl Future<Output = ()> + Send

Wait for the completion of every task in the server.

fn sync_elapsed( &self, ) -> impl Future<Output = Result<Duration, TimestampsError>> + Send

Wait for the completion of every task in the server.

Returns the (approximate) total amount of GPU work done since the last sync.

fn memory_usage(&self) -> MemoryUsage

Get the current memory usage of the server.

fn enable_timestamps(&self)

Enable collecting timestamps.

fn disable_timestamps(&self)

Disable collecting timestamps.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.