cubecl::frontend

Enum TensorArg

pub enum TensorArg<'a, R>
where
    R: Runtime,
{
    Handle {
        handle: TensorHandleRef<'a, R>,
        vectorization_factor: u8,
    },
    Alias {
        input_pos: usize,
    },
}

Argument to be used for tensors passed as arguments to kernels.

Variants

Handle

The tensor is passed with a tensor handle.

Fields

handle: TensorHandleRef<'a, R>

The tensor handle.

vectorization_factor: u8

The vectorization factor.

Alias

The tensor is aliasing another input tensor.

Fields

input_pos: usize

The position of the input tensor.

Implementations

impl<'a, R> TensorArg<'a, R>
where R: Runtime,

pub unsafe fn from_raw_parts<E>( handle: &'a Handle, strides: &'a [usize], shape: &'a [usize], factor: u8, ) -> TensorArg<'a, R>
where E: CubePrimitive,

Create a new tensor argument specified with its vectorization factor.

Safety

If you provide wrong strides or shapes, it might create undefined behavior caused by out-of-bound reads and writes.

pub unsafe fn from_raw_parts_and_size( handle: &'a Handle, strides: &'a [usize], shape: &'a [usize], factor: u8, elem_size: usize, ) -> TensorArg<'a, R>

Create a new tensor argument specified with its vectorization factor with a manual element size in bytes.

Safety

If you provide wrong strides or shapes, it might create undefined behavior caused by out-of-bound reads and writes.

pub fn alias(position: usize) -> TensorArg<'a, R>

Create an alias argument.