cubecl::frontend

Struct Tensor

pub struct Tensor<T>
where
    T: CubeType,
{ /* private fields */ }

The tensor type is similar to the array type, however it comes with more metadata such as stride and shape.

Implementations

impl<T> Tensor<T>
where T: CubeType,

pub fn new() -> Tensor<T>

Constructs a new Tensor.

impl<T> Tensor<T>
where T: CubeType,

pub fn stride<C>(&self, _dim: C) -> u32
where C: Index,

Obtain the stride of input at dimension dim

pub fn shape<C>(&self, _dim: C) -> u32
where C: Index,

Obtain the shape of input at dimension dim

pub fn coordinate<I, D>(&self, _index: I, _dim: D) -> u32
where I: Index, D: Index,

Obtain the coordinate corresponding to the given index of the tensor at dimension dim.

A coordinate is a list of indices corresponding to the multi-dimensional position of an element in the tensor. The dim element in a coordinate is the position along the dim dimension of the tensor.

pub fn len(&self) -> u32

The number of vectorized elements in the tensor.

Warning

The length will be affected by the vectorization factor. To obtain the number of elements, you should multiply the length by the vectorization factor.

pub fn buffer_len(&self) -> u32

The length of the buffer representing the tensor in terms of vectorized elements.

Warning

The buffer length will be affected by the vectorization factor. To obtain the number of elements, you should multiply the length by the vectorization factor.

pub fn rank(&self) -> u32

Returns the rank of the tensor.

pub fn __expand_stride<C>( context: &mut CubeContext, expand: ExpandElementTyped<Tensor<T>>, dim: ExpandElementTyped<u32>, ) -> ExpandElementTyped<u32>
where C: Index,

pub fn __expand_shape<C>( context: &mut CubeContext, expand: ExpandElementTyped<Tensor<T>>, dim: ExpandElementTyped<u32>, ) -> ExpandElementTyped<u32>
where C: Index,

pub fn __expand_coordinate<I, D>( context: &mut CubeContext, expand: ExpandElementTyped<Tensor<T>>, index: ExpandElementTyped<u32>, dim: ExpandElementTyped<u32>, ) -> ExpandElementTyped<u32>
where I: Index, D: Index,

pub fn __expand_len<C>( context: &mut CubeContext, expand: ExpandElementTyped<Tensor<T>>, ) -> ExpandElementTyped<u32>
where C: Index,

pub fn __expand_buffer_len<C>( context: &mut CubeContext, expand: ExpandElementTyped<Tensor<T>>, ) -> ExpandElementTyped<u32>
where C: Index,

pub fn __expand_rank<C>( context: &mut CubeContext, expand: ExpandElementTyped<Tensor<T>>, ) -> ExpandElementTyped<u32>
where C: Index,

impl<E> Tensor<E>
where E: CubePrimitive,

pub unsafe fn index_unchecked<I>(&self, _i: I) -> &E
where I: Index, Tensor<E>: CubeIndex<I>,

Perform an unchecked index into the array

Safety

Out of bounds indexing causes undefined behaviour and may segfault. Ensure index is always in bounds

pub unsafe fn index_assign_unchecked<I>(&mut self, _i: I, _value: E)
where I: Index, Tensor<E>: CubeIndexMut<I>,

Perform an unchecked index assignment into the array

Safety

Out of bounds indexing causes undefined behaviour and may segfault. Ensure index is always in bounds

impl<P> Tensor<Line<P>>
where P: CubePrimitive,

pub fn line_size(&self) -> u32

Get the size of each line contained in the tensor.

Same as the following:

let size = tensor[0].size();

pub fn __expand_line_size( expand: <Tensor<Line<P>> as CubeType>::ExpandType, context: &mut CubeContext, ) -> u32