cubecl::frontend

Trait CubeIndex

pub trait CubeIndex<T>
where
    T: Index,
{
    type Output: CubeType;

    // Provided method
    fn cube_idx(&self, _i: T) -> &Self::Output { ... }
}

Fake indexation so we can rewrite indexes into scalars as calls to this fake function in the non-expanded function

Required Associated Types

type Output: CubeType

Provided Methods

fn cube_idx(&self, _i: T) -> &Self::Output