cubecl::frontend

Trait IntoRuntime

pub trait IntoRuntime: Sized + CubeType {
    // Required method
    fn __expand_runtime_method(
        self,
        context: &mut CubeContext,
    ) -> Self::ExpandType;

    // Provided method
    fn runtime(self) -> Self { ... }
}

Trait useful for cube types that are also used with comptime.

This is used to set a variable as mutable. (Need to be fixed or at least renamed.)

Required Methods

fn __expand_runtime_method(self, context: &mut CubeContext) -> Self::ExpandType

Provided Methods

fn runtime(self) -> Self

Make sure a type is actually expanded into its runtime expand type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.