cubecl

Module prelude

Macros

comment

Insert a literal comment into the kernel source code.

comptime

Mark the contents of this macro as compile time values, turning off all expansion for this code and using it verbatim

Structs

Array

A contiguous array of elements.

ArrayCompilationArg

ArrayHandleRef

Tensor representation with a reference to the server handle.

Atomic

An atomic numerical type wrapping a normal numeric primitive. Enables the use of atomic operations, while disabling normal operations. In WGSL, this is a separate type - on CUDA/SPIR-V it can theoretically be bitcast to a normal number, but this isn’t recommended.

CompiledKernel

A kernel, compiled in the target language

ComputeClient

The ComputeClient is the entry point to require tasks from the ComputeServer. It should be obtained for a specific device via the Compute struct.

CubeContext

CubeDim

ExpandElementTyped

Expand type associated with a type.

FloatExpand

KernelBuilder

Prepare a kernel to create a kernel definition.

KernelDefinition

KernelExpansion

The information necessary to compile a kernel definition.

KernelIntegrator

The kernel integrator allows you to create a kernel definition based on kernel expansion and kernel settings.

KernelLauncher

Prepare a kernel for launch.

KernelSettings

KernelTask

Wraps a kernel to create a cube task.

Line

A contiguous list of elements that supports auto-vectorized operations.

RangeExpand

Registry

It is similar to a map, but where the keys are stored at comptime, but the values can be runtime variables.

ScalarArg

Sequence

A sequence of cube types that is inlined during compilation.

SequenceArg

SequenceCompilationArg

SequenceExpand

Expand type of Sequence.

SharedMemory

Slice

A read-only contiguous list of elements

SliceMut

A read-write contiguous list of elements.

SteppedRangeExpand

SwitchExpand

SwitchExpandExpr

Tensor

The tensor type is similar to the array type, however it comes with more metadata such as stride and shape.

TensorCompilationArg

Compilation argument for a tensor.

TensorHandleRef

Tensor representation with a reference to the server handle, the strides and the shape.

flex32

A floating point type with relaxed precision, minimum f16, max [f32].

tf32

A 19-bit floating point type implementing the tfloat32 format.

Enums

ArrayArg

AtomicOp

CubeCount

Specifieds the number of cubes to be dispatched for a kernel.

ExpandElement

Reference to a JIT variable

IfElseExpand

IfElseExprExpand

TensorArg

Argument to be used for tensors passed as arguments to kernels.

Constants

ABSOLUTE_POS

The position of the working unit in the whole cube kernel, without regards to cubes and axis.

ABSOLUTE_POS_X

The index of the working unit in the whole cube kernel along the X axis, without regards to cubes.

ABSOLUTE_POS_Y

The index of the working unit in the whole cube kernel along the Y axis, without regards to cubes.

ABSOLUTE_POS_Z

The index of the working unit in the whole cube kernel along the Z axis, without regards to cubes.

CUBE_COUNT

The number of cubes launched.

CUBE_COUNT_X

The number of cubes launched along the X axis.

CUBE_COUNT_Y

The number of cubes launched along the Y axis.

CUBE_COUNT_Z

The number of cubes launched along the Z axis.

CUBE_DIM

The total amount of working units in a cube.

CUBE_DIM_X

The dimension of the cube along the X axis.

CUBE_DIM_Y

The dimension of the cube along the Y axis.

CUBE_DIM_Z

The dimension of the cube along the Z axis.

CUBE_POS

The cube position, without regards to axis.

CUBE_POS_X

The cube position along the X axis.

CUBE_POS_Y

The cube position along the Y axis.

CUBE_POS_Z

The cube position along the Z axis.

PLANE_DIM

The total amount of working units in a plane.

UNIT_POS

The position of the working unit inside the cube, without regards to axis.

UNIT_POS_PLANE

The relative position of the working unit inside the plane, without regards to cube dimensions.

UNIT_POS_X

The position of the working unit inside the cube along the X axis.

UNIT_POS_Y

The position of the working unit inside the cube along the Y axis.

UNIT_POS_Z

The position of the working unit inside the cube along the Z axis.

Traits

Abs

ArgSettings

Defines the argument settings used to launch a kernel.

BitCast

Enables reinterpet-casting/bitcasting from any floating point value to any integer value and vice versa

BoolOps

Extension trait for bool.

Cast

Enable elegant casting from any to any CubeElem

Ceil

Clamp

CompilationArg

Argument used during the compilation of kernels.

Cos

CountOnes

CubeElement

The base element trait for the jit backend.

CubeIndex

Fake indexation so we can rewrite indexes into scalars as calls to this fake function in the non-expanded function

CubeIndexMut

CubePrimitive

Form of CubeType that encapsulates all primitive types: Numeric, UInt, Bool

CubeTask

Kernel trait with the ComputeShader that will be compiled and cached based on the provided id.

CubeType

Types used in a cube function must implement this trait

Dot

Erf

Exp

ExpandElementBaseInit

Float

Floating point numbers. Used as input in float kernels

Floor

Index

Init

Trait to be implemented by cube types implementations.

Int

Signed or unsigned integer. Used as input in int kernels

IntoRuntime

Trait useful for cube types that are also used with comptime.

Iterable

Something that can be iterated on by a for loop. Currently only includes Range, StepBy and Sequence.

Kernel

Implement this trait to create a kernel definition.

LaunchArg

Defines a type that can be used as argument to a kernel.

LaunchArgExpand

Defines how a launch argument can be expanded.

Log

Log1p

Magnitude

Max

Min

Normalize

Not

Numeric

Type that encompasses both (unsigned or signed) integers and floats Used in kernels that should work for both.

OptionExt

Powf

Recip

RegistryQuery

To find an item from the registry, the query must be able to be translated to the actual key type.

Remainder

ReverseBits

Round

Runtime

Runtime for the CubeCL.

ScalarArgSettings

Similar to ArgSettings, however only for scalar types that don’t depend on the Runtime trait.

Sin

SizedContainer

SliceOperator

SliceOperatorExpand

Sqrt

Tanh

Functions

array_assign_binary_op_expand

break_expand

copy_bulk

Bulk copy length elements between two array-likes without intermediates.

debug_call_expand

Calls a function and inserts debug symbols if debug is enabled.

debug_source_expand

Adds source instruction if debug is enabled

fma

Fused multiply-add A*B+C.

fma_expand

Expand method of fma.

for_expand

if_else_expand

if_else_expr_expand

if_expand

init_expand

loop_expand

plane_all

Perform a reduce all operation across all units in a plane.

plane_any

Perform a reduce any operation across all units in a plane.

plane_broadcast

Broadcasts the value from the specified plane unit at the given index to all active units within that plane.

plane_elect

Returns true if the cube unit has the lowest plane_unit_id among active unit in the plane

plane_max

Perform a reduce max operation across all units in a plane.

plane_min

Perform a reduce min operation across all units in a plane.

plane_prod

Perform a reduce prod operation across all units in a plane.

plane_sum

Perform a reduce sum operation across all units in a plane.

printf_expand

Prints a formatted message using the print debug layer in Vulkan, or printf in CUDA.

range

integer range. Equivalent to:

range_stepped

Stepped range. Equivalent to:

return_expand

select

Executes both branches, then selects a value based on the condition. This should be branchless, but might depend on the compiler.

select_many

Same as [select] but with lines instead.

slice_expand

spanned_expand

Calls an intrinsic op and inserts debug symbols if debug is enabled.

switch_expand

switch_expand_expr

sync_storage

Sync_storage is the same but change “cube address space(shared memory)” to “storage address space(input args)”. But the set of invocations that are collaborating is still only the invocations in the same cube.There is no guarantee about using barriers alone to make the writes to storage buffer in one cube become visible to invocations in a different cube.

sync_units

Coordinates the following among all invocations in the current cube:

unary_expand

unary_expand_fixed_output

Type Aliases

IntExpand

NumericExpand

RuntimeArg

Runtime arguments to launch a kernel.

Attribute Macros

cube

Mark a cube function, trait or implementation for expansion.

Derive Macros

CubeLaunch

Derive macro to define a cube type that is launched with a kernel

CubeType

Derive macro to define a cube type that is not launched