cubecl::prelude

Struct CubeContext

pub struct CubeContext {
    pub root: Rc<RefCell<Scope>>,
    pub scope: Rc<RefCell<Scope>>,
    pub debug_enabled: bool,
    pub typemap: Rc<RefCell<HashMap<TypeId, Elem>>>,
}

Fields
root: Rc<RefCell<Scope>>scope: Rc<RefCell<Scope>>debug_enabled: booltypemap: Rc<RefCell<HashMap<TypeId, Elem>>>
Implementations

impl CubeContext

pub fn root() -> CubeContext

Create a new cube context, with a root scope A root scope is at the root of a compute shader Therefore there is one cube context per shader The allocator will define the strategy for creating local intermediates and mutable variables

pub fn register<O>(&mut self, op: O)
where O: Into<Instruction>,

pub fn resolve_elem<T>(&self) -> Option<Elem>
where T: 'static,

Resolve the element type of the given generic type.

pub fn register_elem<T>(&mut self, elem: Elem)
where T: 'static,

Register the element type for the given generic type.

pub fn child(&mut self) -> CubeContext

pub fn into_scope(self) -> Scope

pub fn create_local_mut(&mut self, item: Item) -> ExpandElement

Create a new mutable local variable.

pub fn create_local(&mut self, item: Item) -> ExpandElement

Create a new immutable local variable.

pub fn create_local_restricted(&mut self, item: Item) -> ExpandElement

Create a new immutable local binding that must never be a reused variable, regardless of allocator

pub fn create_matrix(&mut self, matrix: Matrix) -> ExpandElement

Create a new matrix element.

pub fn create_slice(&mut self, item: Item) -> ExpandElement

Create a new slice element.

pub fn create_shared(&mut self, item: Item, size: u32) -> ExpandElement

pub fn create_local_array(&mut self, item: Item, size: u32) -> ExpandElement

pub fn create_const_array( &mut self, item: Item, data: Vec<Variable>, ) -> ExpandElement

pub fn input(&mut self, id: u32, item: Item) -> ExpandElement

Obtain the index-th input

pub fn output(&mut self, id: u32, item: Item) -> ExpandElement

Obtain the index-th output

pub fn scalar(&self, id: u32, elem: Elem) -> ExpandElement

Obtain the index-th scalar