cubecl::prelude

Struct SliceMut

pub struct SliceMut<E> { /* private fields */ }

A read-write contiguous list of elements.

Safety

Since data can be accessed by any unit during kernel execution, this can never be safe.

Implementations

impl<E> SliceMut<E>

pub fn len(&self) -> u32

Get the length of the slice.

pub fn into_aligned(self) -> SliceMut<Line<E>>
where E: CubePrimitive,

Returns the same slice, but with lines of length 1.

impl<E> SliceMut<E>
where E: CubePrimitive,

pub unsafe fn index_unchecked<I>(&self, _i: I) -> &E
where I: Index, SliceMut<E>: CubeIndex<I>,

Perform an unchecked index into the array

Safety

Out of bounds indexing causes undefined behaviour and may segfault. Ensure index is always in bounds

pub unsafe fn index_assign_unchecked<I>(&mut self, _i: I, _value: E)
where I: Index, SliceMut<E>: CubeIndexMut<I>,

Perform an unchecked index assignment into the array

Safety

Out of bounds indexing causes undefined behaviour and may segfault. Ensure index is always in bounds