cubecl::prelude

Trait RegistryQuery

pub trait RegistryQuery<K>: Into<K> { }

To find an item from the registry, the query must be able to be translated to the actual key type.

Example

If you use u32 as key that may become ExpandElementTyped during the expansion, both types need to implement RegistryQuery.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.