cubecl::frontend

Trait SliceOperator

pub trait SliceOperator<E>: CubeType<ExpandType = Self::Expand>
where
    E: CubeType,
{
    type Expand: SliceOperatorExpand<E>;

    // Provided methods
    fn slice<Start, End>(&self, start: Start, end: End) -> Slice<E> 
       where Start: Index,
             End: Index { ... }
    fn __expand_slice(
        context: &mut CubeContext,
        expand: Self::Expand,
        start: ExpandElementTyped<u32>,
        end: ExpandElementTyped<u32>,
    ) -> ExpandElementTyped<Slice<E>> { ... }
    fn slice_mut<Start, End>(&mut self, start: Start, end: End) -> SliceMut<E>
       where Start: Index,
             End: Index { ... }
    fn __expand_slice_mut(
        context: &mut CubeContext,
        expand: Self::Expand,
        start: ExpandElementTyped<u32>,
        end: ExpandElementTyped<u32>,
    ) -> ExpandElementTyped<SliceMut<E>> { ... }
    fn to_slice(&self) -> Slice<E>  { ... }
    fn __expand_to_slice(
        context: &mut CubeContext,
        expand: Self::Expand,
    ) -> ExpandElementTyped<Slice<E>> { ... }
    fn to_slice_mut(&mut self) -> SliceMut<E> { ... }
    fn __expand_to_slice_mut(
        context: &mut CubeContext,
        expand: Self::Expand,
    ) -> ExpandElementTyped<SliceMut<E>> { ... }
}

Required Associated Types

type Expand: SliceOperatorExpand<E>

Provided Methods

fn slice<Start, End>(&self, start: Start, end: End) -> Slice<E>
where Start: Index, End: Index,

Return a read-only view of all elements comprise between the start and end indices. In checked mode, if the end index is out-of-bound, it is replaced by the length of self.

fn __expand_slice( context: &mut CubeContext, expand: Self::Expand, start: ExpandElementTyped<u32>, end: ExpandElementTyped<u32>, ) -> ExpandElementTyped<Slice<E>>

Expand function of SliceOperator::slice.

fn slice_mut<Start, End>(&mut self, start: Start, end: End) -> SliceMut<E>
where Start: Index, End: Index,

Return a read-write view of all elements comprise between the start and end indices. In checked mode, if the end index is out-of-bound, it is replaced by the length of self.

fn __expand_slice_mut( context: &mut CubeContext, expand: Self::Expand, start: ExpandElementTyped<u32>, end: ExpandElementTyped<u32>, ) -> ExpandElementTyped<SliceMut<E>>

Expand function of SliceOperator::slice_mut.

fn to_slice(&self) -> Slice<E>

Reinterprete the current type as a read-only slice.

fn __expand_to_slice( context: &mut CubeContext, expand: Self::Expand, ) -> ExpandElementTyped<Slice<E>>

Expand function of SliceOperator::to_slice.

fn to_slice_mut(&mut self) -> SliceMut<E>

Reinterprete the current type as a read-write slice.

fn __expand_to_slice_mut( context: &mut CubeContext, expand: Self::Expand, ) -> ExpandElementTyped<SliceMut<E>>

Expand function of SliceOperator::to_slice_mut.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.